requirements.txt
selenium
webdriver-manager
firebase-admin
python-dotenv
pandas
schedule
